install flatpaks only on desktop

+++ .local/share/bin/init-user-env.sh
@@ -31,16 +31,20 @@ for path in "${!repos[@]}"; do
 done
 
 flatpak remote-add --user --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
-flatpak install -y --user flathub org.telegram.desktop || true
-flatpak override --user org.telegram.desktop --filesystem="${HOME}/Pictures"
-flatpak override --user org.telegram.desktop --filesystem="${HOME}/Downloads"
 
-flatpak install -y --user org.chromium.Chromium
-flatpak install -y --user org.gnome.Evince
-flatpak install -y --user org.keepassxc.KeePassXC
-flatpak install -y --user com.parsecgaming.parsec
-flatpak install -y --user flathub org.libreoffice.LibreOffice
-flatpak install -y --user flathub org.wezfurlong.wezterm
+if dpkg -l | grep ubuntu-desktop
+then
+    flatpak install -y --user flathub org.telegram.desktop || true
+    flatpak override --user org.telegram.desktop --filesystem="${HOME}/Pictures"
+    flatpak override --user org.telegram.desktop --filesystem="${HOME}/Downloads"
+
+    flatpak install -y --user org.chromium.Chromium
+    flatpak install -y --user org.gnome.Evince
+    flatpak install -y --user org.keepassxc.KeePassXC
+    flatpak install -y --user com.parsecgaming.parsec
+    flatpak install -y --user flathub org.libreoffice.LibreOffice
+    flatpak install -y --user flathub org.wezfurlong.wezterm
+fi
 
 nix-channel --add https://nixos.org/channels/nixos-24.05 nixpkgs
 nix-channel --update
